Add unit tests for DiffTracker functionality in service gateway
- Implement tests for Kubernetes state updates in k8s_state_updates_test.go, covering service operations, endpoint updates, pod updates, and identity changes.
- Add tests for NRP state updates in nrp_state_updates_test.go, focusing on load balancers, NAT gateways, and location address synchronization.
- Create sync operation tests in sync_operations_test.go to validate the synchronization logic between Kubernetes and NRP resources.
- Enhance utility tests in util_test.go to verify deep equality checks and operation string representations.
